Spirit airlines and Expedia websites being scraped and saved
Implemented algorithm for the Spirit airlines site that chooses the cheapest flight uses the values associated with that prices to send an email to the specified email address. Currently the time between runs is 15 seconds. Updates will be running every hour for 12 or 24 hours and saving the trends into an excel sheet.